+++ isitfit/__init__.py
@@ -1,1 +1,1 @@
-isitfit_version='0.19.5'
+isitfit_version='0.19.6rc'